requirements.txt
discord.py==1.5.1
peewee==3.14.0
python-dotenv==0.15.0
